+++ src/interfaces/discord/commands/discord-commands.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { DiscordConfigureServerCommand } from './discord-configure-server-command.service';
-import { DiscordAttachmentService } from '../services/chats/discord-attachment.service';
+import { DiscordConfigureServerCommand } from './server-config-command/discord-configure-server.command';
+import { DiscordAttachmentService } from '../services/chat-command/discord-attachment.service';
 import { OpenaiModule } from '../../../core/ai/openai/openai.module';
 
 @Module({
