+++ src/interfaces/discord/discord.module.ts
@@ -4,18 +4,18 @@ import { NecordModule } from 'necord';
 
 import { DiscordConfigService } from './services/discord-config.service';
 import { DiscordEventService } from './services/events/discord-event.service';
-import { DiscordChatService } from './services/chats/discord-chat.service';
+import { DiscordChatService } from './services/chat-command/discord-chat.service';
 import { GcpModule } from '../../core/gcp/gcp.module';
 import { UsersModule } from '../../core/database/collections/users/users.module';
 import { ChatbotModule } from '../../core/chatbot/chatbot.module';
 import { AiChatsModule } from '../../core/database/collections/ai-chats/ai-chats.module';
 import { CoupletsModule } from '../../core/database/collections/couplets/couplets.module';
 import { MessagesModule } from '../../core/database/collections/messages/messages.module';
-import { DiscordContextService } from './services/chats/discord-context.service';
+import { DiscordContextService } from './services/chat-command/discord-context.service';
 import { DiscordMongodbService } from './services/discord-mongodb.service';
-import { DiscordMessageService } from './services/chats/discord-message.service';
+import { DiscordMessageService } from './services/chat-command/discord-message.service';
 import { OpenaiModule } from '../../core/ai/openai/openai.module';
-import { DiscordAttachmentService } from './services/chats/discord-attachment.service';
+import { DiscordAttachmentService } from './services/chat-command/discord-attachment.service';
 import { DiscordOnMessageService } from './services/events/discord-on-message.service';
 import { DiscordCommandsModule } from './commands/discord-commands.module';
