+++ src/interfaces/discord/services/chat-command/discord-attachment.service.ts
