+++ src/interfaces/discord/services/events/discord-on-message.service.ts
@@ -1,10 +1,10 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { AiChatsService } from '../../../../core/database/collections/ai-chats/ai-chats.service';
 import { Message, ThreadChannel } from 'discord.js';
-import { DiscordMessageService } from '../chats/discord-message.service';
+import { DiscordMessageService } from '../chat-command/discord-message.service';
 import { ChatbotManagerService } from '../../../../core/chatbot/chatbot-manager.service';
 import { AiChatDocument } from '../../../../core/database/collections/ai-chats/ai-chat.schema';
-import { DiscordContextService } from '../chats/discord-context.service';
+import { DiscordContextService } from '../chat-command/discord-context.service';
 import { UsersService } from '../../../../core/database/collections/users/users.service';
 import { OpenaiChatService } from '../../../../core/ai/openai/openai-chat.service';
